Restructure project into three core modules
