+++ Procfile
@@ -1,0 +1,1 @@
+worker: python molly.py
