+++ requirements.txt
@@ -1,0 +1,4 @@
+python-telegram-bot==20.7
+python-dotenv
+torch
+numpy
